Add exact geometric predicates using simulation of simplicity
In order to avoid the complexities and costs of floating point expansion
arithmetic, our predicates require the user to quantize into integer
space (see the untested quantize.h file).  We can still filter using
interval arithmetic, but if that fails we can now switch to compile-time
precision integer arithmetic to handle all degenerate cases.

The simplicity of simplicity code generator uses the Sage computer
algebra system.  As a consequence the code generator itself is GPL,
but the code it generates and surrounding helper utilities are BSD
like the rest of core.