exact/scope: Trick clang into not messing with our rounding modes
The test_circles unit test is failing in release mode but not in debug
mode, and the differences appear to be due to clang moving arithmetic
across a call to fesetround.  This is now fixed with a hack: we assert
that the fesetround is successful, which apparently stops clang's mode
motion.  This is not a perfect fix, so we may see similar problems in
future.