metabuild: Lock -march to corei7, and turn off -Werror (for now)
For some reason, geode generates a ton of warnings. By default on windows it
builds with -Werror, which causes everything to explode.

Additionally, and along the same lines of #omco/mpir/1, geode defaults to the
'native' architecture, which can result in GCC generating assembly that is
invalid for CPUs other than the one geode was built on. This can cause Otherplan
to crash due to an illegal instruction. This patch updates metaBuild.sh to use
the corei7 architecture, which is the same specified in Otherplan's qmake
configuration.