Hacks to make things work on Windows
Now the pyd should not be copied into geode/geode; it should be left
where it is in build/native/release/lib. This location was added to
PYTHONPATH, and most of the __init__.py files were changed to reflect
this.

This commit was made while working with the combined geode/tubecad
codebase. However it should work just with geode.