Add WIP sketch with idea for profiling specific blocks
Idea is to get a call tree with some basic timing info in a more or less in-band way that could be used for automated tests. Sort of like Instruments or Gprof but without running an external application or changing the full build process. Would require explicitly marking regions of interest, but avoids wading through benchmarks of everything else.